Refactor turnkey files
